requirements.txt
gymnasium
flappy-bird-gymnasium
torch
numpy
matplotlib
PyYAML
